# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Mailflare v1 API
  version: "1.0.0"
  description: |
    The public, agent-facing surface of Mailflare (`/api/v1/**`).

    Every route authenticates with an API key (`Authorization: Bearer ep_...`)
    and needs exactly one scope. Keys are issued in the dashboard under
    **API Keys**, or through `POST /api/api-keys` with a cookie session.

    - **Tenancy.** A key acts only inside the organisation it was issued in, and
      only on the mailboxes its owner may open. Anything else reads as missing.
    - **Errors.** Always `{ "error": "...", "code": "..." }`; `code` is a stable
      machine-readable tag and may be absent on responses produced by the
      authentication layer.
    - **Paging.** Cursor-based. Pass the `nextCursor` of a page back as
      `cursor`; `nextCursor: null` means the last page.
    - **Rate limit.** 600 requests per minute per key. Over that, `429` with a
      `Retry-After` header in seconds.
  license:
    name: MIT
servers:
  - url: https://mail.easyserver.net/api
    description: Production
  - url: http://localhost:3000/api
    description: Local development
security:
  - apiKey: []
...
paths:
...
  /v1/messages/{id}:
    patch:
      tags: [Messages]
      summary: Update a message
      description: |
        `read` and `starred` need `read_only` on the mailbox; `status` and
        `snoozedUntil` need `full_access`. Scope: `messages:write`.
      operationId: updateMessage
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MessagePatch" }
      responses:
        "200":
          description: The updated fields.
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { $ref: "#/components/schemas/MessageState" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: ep_...
      description: |
        `Authorization: Bearer ep_...`. Session cookies are never accepted here,
        and a key is never accepted on the dashboard routes.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, maxLength: 200 }
...
  responses:
    BadRequest:
      description: The request body or query string did not validate.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing, unknown, revoked or expired key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            revoked: { value: { error: "API key revoked" } }
            expired: { value: { error: "API key expired" } }
            unknown: { value: { error: "Unauthorized" } }
    Forbidden:
      description: |
        The key lacks the route's scope, its organisation is suspended, or its
        owner lacks the required rank on the mailbox.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            scope: { value: { error: "Insufficient scope" } }
            rank: { value: { error: "Forbidden", code: "forbidden" } }
    NotFound:
      description: |
        No such row inside this organisation, or one in a mailbox the key's
        owner cannot read — the two are deliberately indistinguishable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PayloadTooLarge:
      description: The request body exceeded the route's limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: More than 600 requests in a minute on this key.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: "Rate limit exceeded", code: "rate_limited" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code:
          type: string
          description: Stable machine-readable tag; absent on authentication errors.
          enum:
            - invalid_body
            - invalid_query
            - invalid_assignee
            - no_recipient
            - not_found
            - forbidden
            - payload_too_large
            - rate_limited
            - send_failed
...
    MessageStatus:
      type: string
      enum: [received, sent, draft, trash, spam]
...
    MessageState:
      type: object
      required: [id, read, starred, status]
      properties:
        id: { type: string }
        mailboxId: { type: [string, "null"] }
        conversationId: { type: [string, "null"] }
        read: { type: boolean }
        starred: { type: boolean }
        status: { $ref: "#/components/schemas/MessageStatus" }
        snoozedUntil: { type: [string, "null"], format: date-time }
    MessagePatch:
      type: object
      minProperties: 1
      properties:
        read: { type: boolean }
        starred: { type: boolean }
        status: { $ref: "#/components/schemas/MessageStatus" }
        snoozedUntil:
          type: [string, "null"]
          format: date-time
          description: An ISO timestamp in the future, or `null` to clear the snooze.
...